requirements.txt
attrs==19.3.0
aws==0.2.5
awscli==1.16.278
bcrypt==3.1.7
boto==2.49.0
boto3==1.10.14
botocore==1.13.14
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
CodeConvert==3.0.2
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.6
django-base-url==2019.7.24
django-models-ext==1.1.9
django-rest-swagger==2.1.2
django-six==1.0.4
django-storages==1.8
djangorestframework==3.10.3
docutils==0.15.2
fabric==2.5.0
furl==2.1.0
idna==2.8
importlib-metadata==1.1.0
invoke==1.3.0
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
more-itertools==8.0.0
oauthlib==3.1.0
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==19.2
paramiko==2.6.0
pluggy==0.13.1
prettytable==0.7.2
py==1.8.0
pyasn1==0.4.7
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.2.1
shortuuid==0.5.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
TimeConvert==1.5.0
tzlocal==2.0.0
unpyclib==0.8.1
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
xlwt==1.3.0
zipp==0.6.0
